Use __init__.py for top-level imports only
